v0.3.2
N: set opcache.max_accelerated_files=10000 and opcache.revalidate_freq=1
to prevent application warnings
N: add button 'Tune Cache' to activate
the APCu memcache for Next/ownCloud
F: create link to certificate
/usr/local/etc/ssl/cert.pem if not already exists to prevent application
updater problems
C: update translations base file on Launchpad
N:
Kurdish translation
C: update German translation
C: update Greek
translation
C: update Turkish translation

+++ owncloud/owncloud-start.php
@@ -33,8 +33,6 @@ if (is_link("/usr/local/www/owncloud-config.php")) unlink("/usr/local/www/ownclo
 if (is_link("/usr/local/www/owncloud-update_extension.php")) unlink("/usr/local/www/owncloud-update_extension.php");
 if (is_link("/usr/local/www/ext/owncloud")) unlink("/usr/local/www/ext/owncloud");
 mwexec("rmdir -p /usr/local/www/ext");
-if (is_link("/usr/local/share/certs/ca-root-nss.crt")) unlink("/usr/local/share/certs/ca-root-nss.crt");
-mwexec("rmdir -p /usr/local/share/certs");
 $return_val = 0; 
 // create links to extension files
 $return_val += mwexec("ln -sw {$rootfolder}/locale-owncloud /usr/local/share/", true);
@@ -42,10 +40,14 @@ $return_val += mwexec("ln -sw {$rootfolder}/owncloud-config.php /usr/local/www/o
 $return_val += mwexec("ln -sw {$rootfolder}/owncloud-update_extension.php /usr/local/www/owncloud-update_extension.php", true);
 $return_val += mwexec("mkdir -p /usr/local/www/ext", true);
 $return_val += mwexec("ln -sw {$rootfolder}/owncloud /usr/local/www/ext/owncloud", true);
-$return_val += mwexec("echo upload_tmp_dir = {$config['websrv']['uploaddir']} > /usr/local/etc/php/nextowncloud-php.ini && service websrv restart", true);
+// use z-... prefix to override system defaults => opcache.max_accelerated_files=10000 and opcache.revalidate_freq=1
+$return_val += mwexec("cp {$rootfolder}/z-nextowncloud-php.ini /usr/local/etc/php/", true);
+$return_val += mwexec("echo upload_tmp_dir={$config['websrv']['uploaddir']} >> /usr/local/etc/php/z-nextowncloud-php.ini && service websrv restart", true);
 // required for updater to work
-$return_val += mwexec("mkdir -p /usr/local/share/certs", true);
-$return_val += mwexec("ln -sw /usr/local/etc/ssl/cert.pem /usr/local/share/certs/ca-root-nss.crt", true);
+if (!is_file("/usr/local/share/certs/ca-root-nss.crt")) {
+	$return_val += mwexec("mkdir -p /usr/local/share/certs", true);
+	$return_val += mwexec("ln -sw /usr/local/etc/ssl/cert.pem /usr/local/share/certs/ca-root-nss.crt", true);
+}
 if ($return_val == 0) mwexec("logger nextowncloud-extension: GUI loaded");
 else mwexec("logger nextowncloud-extension: error(s) during startup, failed with return value = {$return_val}"); 
 ?>
